+++ fiddles/angular2-cli/fiddle-0011-CssAnimation/src/app/components/loading-anime/index.ts
@@ -1,0 +1,2 @@
+export * from './loading-anime-options';
+export * from './loading-anime.component';