+++ fiddles/angular2-cli/fiddle-0011-CssAnimation/src/app/components/loading-anime/loading-anime-options.ts
@@ -1,0 +1,7 @@
+export interface LoadingAnimeOptions {
+  isDataValid?:boolean;   // validation flag that enables message text
+  error?:string;          // error display message
+  warning?:string;        // warning display message
+  isLoading?:boolean;     // anime on/off flag
+  cssClass?:string;       // template css class override
+}