+++ fiddles/angular2-cli/fiddle-0011-CssAnimation/src/app/components/loading-anime/loading-anime.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { LoadingAnimeComponent } from './loading-anime.component';
+
+describe('LoadingAnimeComponent', () => {
+  let component: LoadingAnimeComponent;
+  let fixture: ComponentFixture<LoadingAnimeComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ LoadingAnimeComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(LoadingAnimeComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should be created', () => {
+    expect(component).toBeTruthy();
+  });
+});