+++ fiddles/nativeScript/fiddle-0012-LocalStorage/pagenavigation/app/app.js
@@ -1,0 +1,3 @@
+var application = require("application");
+require('nativescript-localstorage');
+application.start({ moduleName: "main-page" });