+++ fiddles/nativeScript/fiddle-0012-LocalStorage/pagenavigation/app/main-view-model.js
@@ -1,0 +1,30 @@
+var Observable = require("data/observable").Observable;
+var frame = require('ui/frame');
+var localStorage = require('nativescript-localstorage');
+var Observable = require("data/observable").Observable;
+
+function getMessage(counter) {
+  if (counter <= 0) {
+    localStorage.clear();
+    console.log("Length", localStorage.length);
+    return "Hoorraaay! You unlocked the NativeScript clicker achievement!";
+  } else {
+    console.log("Length", localStorage.length);
+    return counter + " taps left";
+  }
+}
+
+function createViewModel() {
+  var viewModel = new Observable();
+  viewModel.counter = localStorage.getItem("counter") || 42;
+  viewModel.message = getMessage(viewModel.counter);
+
+  viewModel.onTap = function() {
+    this.counter--;
+    localStorage.setItem("counter", this.counter);
+    this.set("message", getMessage(this.counter));
+  }
+
+  return viewModel;
+}
+exports.createViewModel = createViewModel;